Add support for XinJE TCP protocol.
This is an undocumented protocol, which functions through Modbus. It
uses function codes 30-33 to read and write registers and coils, has
32-bit addresses, and does not distinguish r/w from ro registers.

Currently, all registers are treated as 4 byte words. This is *not* a
restriction of the protocol, but is only implemented for convenience.
A better implementation would read registers as 2 byte words, similarly
to standard Modbus functions. This leaves the potential for future work.